+++ TestSipParser/TestSipFrom.cpp
@@ -134,7 +134,7 @@ bool TestSipFrom()
 	const char * pszTest = "<sip:1001@192.168.0.2;x-nearend;x-refci=26528012;x-nearendclusterid=StandAloneCluster;x-nearenddevice=SEP002545941441;x-nearendaddr=1001;x-farendrefci=26528013;x-farendclusterid=StandAloneCluster;x-farenddevice=SEPF02929E28A35;x-farendaddr=1000>;tag=84501~ec4d89fd-07f4-46f8-99c5-63220364acb6-26528017";
 	std::string strValue;
 
-	clsFrom.Parse( pszTest, strlen(pszTest) );
+	clsFrom.Parse( pszTest, (int)strlen(pszTest) );
 
 	SearchSipParameter( clsFrom.m_clsUri.m_clsUriParamList, "x-farendaddr", strValue );
